refactor: split parser
