+++ benches/generate_events.rs
@@ -6,7 +6,7 @@ use criterion::BenchmarkId;
 use criterion::Criterion;
 use nexmark_server::create_generators_for_config;
 use nexmark_server::generator::source::NexmarkSource;
-use nexmark_server::NexmarkConfig;
+use nexmark_server::parser::NexmarkConfig;
 use rand_chacha::ChaCha8Rng;
 
 fn event_generation(c: &mut Criterion) {
@@ -14,7 +14,7 @@ fn event_generation(c: &mut Criterion) {
     // vary the qps and monitor the time required to send 1,000 events
     for qps in [1_000, 10_000, 100_000, 1_000_000].iter() {
         let mut nexmark_config = NexmarkConfig::default();
-        nexmark_config.first_event_rate = *qps;
+        nexmark_config.event_rate = *qps;
         let nexmark_source = Arc::new(NexmarkSource::new(&nexmark_config));
         group.bench_with_input(BenchmarkId::from_parameter(qps), qps, |b, &_qps| {
             b.to_async(tokio::runtime::Runtime::new().unwrap())
@@ -28,7 +28,7 @@ fn event_generation(c: &mut Criterion) {
     for num_gen in 1..10 {
         let mut nexmark_config = NexmarkConfig::default();
         nexmark_config.num_event_generators = num_gen as usize;
-        nexmark_config.first_event_rate = 10_000;
+        nexmark_config.event_rate = 10_000;
         nexmark_config.max_events = 1_000;
         let nexmark_source = Arc::new(NexmarkSource::new(&nexmark_config));
         group.bench_with_input(
@@ -46,11 +46,11 @@ fn event_generation(c: &mut Criterion) {
     // since the qps is 10_000, gold standard is 0.1s per iter
     for byte_size in [100, 200, 300, 400, 500] {
         let mut nexmark_config = NexmarkConfig::default();
-        nexmark_config.first_event_rate = 10_000;
+        nexmark_config.event_rate = 10_000;
         nexmark_config.max_events = 1_000;
-        nexmark_config.avg_auction_byte_size = byte_size;
-        nexmark_config.avg_person_byte_size = byte_size;
-        nexmark_config.avg_bid_byte_size = byte_size;
+        nexmark_config.additional_auction_byte_size = byte_size;
+        nexmark_config.additional_bid_byte_size = byte_size;
+        nexmark_config.additional_person_byte_size = byte_size;
         let nexmark_source = Arc::new(NexmarkSource::new(&nexmark_config));
         group.bench_with_input(
             BenchmarkId::from_parameter(byte_size),
